Move tqdm_gui into its own file, exclude from coverage
