+++ .coveragerc
@@ -2,3 +2,4 @@
 branch = True
 omit =
     tqdm/tests/*
+    tqdm/_tqdm_gui.py
