+++ tqdm/__init__.py
@@ -1,9 +1,10 @@
 from ._tqdm import tqdm
-from ._tqdm import tqdm_gui
 from ._tqdm import trange
 from ._tqdm import format_interval
 from ._tqdm import format_meter
+from ._tqdm_gui import tqdm_gui
+from ._tqdm_gui import tgrange
 from ._version import __version__  # NOQA
 
-__all__ = ['tqdm', 'tqdm_gui', 'trange',
+__all__ = ['tqdm', 'tqdm_gui', 'trange', 'tgrange',
            'format_interval', 'format_meter', '__version__']
